api/requirements.txt
uvicorn==0.23.2
fastapi==0.101.1
python-dotenv==1.0.0
